+++ app/(dashboard)/components/InspectionStatus/index.tsx
