+++ app/(dashboard)/components/InspectionStatus/styles.css.ts
