+++ app/(dashboard)/components/NoticeBoard/styles.css.ts
