+++ app/(dashboard)/components/VehicleStatus/styles.css.ts
