+++ app/history/styles.css.ts
