+++ app/location/components/VehicleStatus/index.tsx
