+++ app/route/page.tsx
@@ -3,8 +3,8 @@
 import { useEffect, useState } from 'react'
 import { Polyline } from 'react-kakao-maps-sdk'
 
-import Map from '@/app/(dashboard)/location/components/Map'
-import RouteSearchPanel from '@/app/(dashboard)/route/components/RouteSearchPanel'
+import Map from '@/app/location/components/Map'
+import RouteSearchPanel from '@/app/route/components/RouteSearchPanel'
 import mockRoutesData from '@/mock/vehicle_route_data.json'
 import { styles } from '@/styles/theme.css'
 import { PathPoint } from '@/types/location'
